contra-youtube-recommender/requirements.txt
# Core dependencies
google-api-python-client==2.111.0  # YouTube Data API
google-auth-oauthlib==1.2.0        # OAuth authentication
google-auth-httplib2==0.2.0

# Embedding and ML
sentence-transformers==2.3.1        # Open-source embeddings
torch==2.1.2                        # PyTorch for embeddings
numpy==1.26.3                       # Numerical operations
scikit-learn==1.4.0                 # Distance calculations

# Data processing
pandas==2.2.0                       # Data manipulation
tqdm==4.66.1                        # Progress bars

# Web framework
fastapi==0.109.0                    # API framework
uvicorn==0.27.0                     # ASGI server
pydantic==2.5.3                     # Data validation

# Database
sqlalchemy==2.0.25                  # ORM
alembic==1.13.1                     # Database migrations

# Utilities
python-dotenv==1.0.0                # Environment variables
requests==2.31.0                    # HTTP requests
aiohttp==3.9.1                      # Async HTTP

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
